requirements.txt
fastapi
uvicorn[standard]
aiogram>=3.0.0
sqlalchemy[asyncio]>=2.0.0
alembic
asyncpg
pydantic-settings
structlog
python-dotenv
greenlet
pytest
pytest-asyncio
httpx
aiosqlite
